+++ src/main/java/com/ing/data/cassandra/jdbc/CassandraPreparedStatement.java
@@ -45,6 +45,7 @@ import java.net.Inet4Address;
 import java.net.InetAddress;
 import java.net.URL;
 import java.nio.ByteBuffer;
+import java.sql.BatchUpdateException;
 import java.sql.Blob;
 import java.sql.Clob;
 import java.sql.Date;
@@ -78,6 +79,8 @@ import static com.ing.data.cassandra.jdbc.utils.ConversionsUtil.convertToByteArr
 import static com.ing.data.cassandra.jdbc.utils.ConversionsUtil.convertToInstant;
 import static com.ing.data.cassandra.jdbc.utils.ConversionsUtil.convertToLocalDate;
 import static com.ing.data.cassandra.jdbc.utils.ConversionsUtil.convertToLocalTime;
+import static com.ing.data.cassandra.jdbc.utils.ErrorConstants.BATCH_STATEMENT_FAILURE_MSG;
+import static com.ing.data.cassandra.jdbc.utils.ErrorConstants.BATCH_UPDATE_FAILED;
 import static com.ing.data.cassandra.jdbc.utils.ErrorConstants.MUST_BE_POSITIVE_BINDING_INDEX;
 import static com.ing.data.cassandra.jdbc.utils.ErrorConstants.NO_RESULT_SET;
 import static com.ing.data.cassandra.jdbc.utils.ErrorConstants.OUT_OF_BOUNDS_BINDING_INDEX;
@@ -277,18 +280,41 @@ public class CassandraPreparedStatement extends CassandraStatement
             }
 
             int i = 0;
+            boolean hasFailures = false;
+            final StringBuilder errMsgBuilder = new StringBuilder(BATCH_UPDATE_FAILED);
             for (final CompletionStage<AsyncResultSet> future : futures) {
-                CompletableFutures.getUninterruptibly(future);
-                returnCounts[i] = 1;
-                i++;
+                try {
+                    final AsyncResultSet asyncResultSet = CompletableFutures.getUninterruptibly(future);
+                    if (asyncResultSet.getColumnDefinitions().size() > 0) {
+                        returnCounts[i] = EXECUTE_FAILED;
+                        hasFailures = true;
+                        errMsgBuilder.append(String.format(BATCH_STATEMENT_FAILURE_MSG, i,
+                            "attempts to return a result set"));
+                    } else {
+                        returnCounts[i] = SUCCESS_NO_INFO;
+                    }
+                } catch (final Exception e) {
+                    returnCounts[i] = EXECUTE_FAILED;
+                    hasFailures = true;
+                    errMsgBuilder.append(String.format(BATCH_STATEMENT_FAILURE_MSG, i, e.getMessage()));
+                } finally {
+                    i++;
+                }
             }
 
-            // Empty the batch statement list after execution.
-            this.batchStatements = new ArrayList<>();
+            // If at least one statement in the batch has failed, throw a BatchUpdateException with the appropriate
+            // updateCounts (i.e. failed statements have value EXECUTE_FAILED).
+            if (hasFailures) {
+                throw new BatchUpdateException(errMsgBuilder.toString(), returnCounts);
+            }
+        } catch (final SQLException sqlEx) {
+            // Any SQLException can be re-throw as is, without being wrapped into another SQLException.
+            throw sqlEx;
         } catch (final Exception e) {
+            throw new SQLTransientException(e);
+        } finally {
             // Empty the batch statement list after execution even if it failed.
             this.batchStatements = new ArrayList<>();
-            throw new SQLTransientException(e);
         }
 
         return returnCounts;