+++ src/main/java/com/ing/data/cassandra/jdbc/CassandraStatement.java
@@ -28,6 +28,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.annotation.Nonnull;
+import java.sql.BatchUpdateException;
 import java.sql.CallableStatement;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -54,6 +55,8 @@ import static com.ing.data.cassandra.jdbc.utils.ErrorConstants.BAD_FETCH_SIZE;
 import static com.ing.data.cassandra.jdbc.utils.ErrorConstants.BAD_HOLD_RS;
 import static com.ing.data.cassandra.jdbc.utils.ErrorConstants.BAD_KEEP_RS;
 import static com.ing.data.cassandra.jdbc.utils.ErrorConstants.BAD_TYPE_RS;
+import static com.ing.data.cassandra.jdbc.utils.ErrorConstants.BATCH_STATEMENT_FAILURE_MSG;
+import static com.ing.data.cassandra.jdbc.utils.ErrorConstants.BATCH_UPDATE_FAILED;
 import static com.ing.data.cassandra.jdbc.utils.ErrorConstants.NO_GEN_KEYS;
 import static com.ing.data.cassandra.jdbc.utils.ErrorConstants.NO_MULTIPLE;
 import static com.ing.data.cassandra.jdbc.utils.ErrorConstants.NO_RESULT_SET;
@@ -426,30 +429,62 @@ public class CassandraStatement extends AbstractStatement
     @Override
     public int[] executeBatch() throws SQLException {
         final int[] returnCounts = new int[this.batchQueries.size()];
-        final List<CompletionStage<AsyncResultSet>> futures = new ArrayList<>();
-        if (LOG.isTraceEnabled() || this.connection.isDebugMode()) {
-            LOG.debug("CQL statements: {}", this.batchQueries.size());
-        }
-
-        for (final String query : this.batchQueries) {
+        try {
+            final List<CompletionStage<AsyncResultSet>> futures = new ArrayList<>();
             if (LOG.isTraceEnabled() || this.connection.isDebugMode()) {
-                LOG.debug("CQL: {}", query);
+                LOG.debug("CQL statements: {}", this.batchQueries.size());
             }
-            SimpleStatement stmt = SimpleStatement.newInstance(query)
-                .setConsistencyLevel(this.connection.getDefaultConsistencyLevel());
-            if (this.customTimeoutProfile != null) {
-                stmt = stmt.setExecutionProfile(this.customTimeoutProfile);
-            }
-            final CompletionStage<AsyncResultSet> resultSetFuture =
-                ((CqlSession) this.connection.getSession()).executeAsync(stmt);
-            futures.add(resultSetFuture);
-        }
 
-        int i = 0;
-        for (final CompletionStage<AsyncResultSet> future : futures) {
-            CompletableFutures.getUninterruptibly(future);
-            returnCounts[i] = 1;
-            i++;
+            for (final String query : this.batchQueries) {
+                if (LOG.isTraceEnabled() || this.connection.isDebugMode()) {
+                    LOG.debug("CQL: {}", query);
+                }
+                SimpleStatement stmt = SimpleStatement.newInstance(query)
+                    .setConsistencyLevel(this.connection.getDefaultConsistencyLevel());
+                if (this.customTimeoutProfile != null) {
+                    stmt = stmt.setExecutionProfile(this.customTimeoutProfile);
+                }
+                final CompletionStage<AsyncResultSet> resultSetFuture =
+                    ((CqlSession) this.connection.getSession()).executeAsync(stmt);
+                futures.add(resultSetFuture);
+            }
+
+            int i = 0;
+            boolean hasFailures = false;
+            final StringBuilder errMsgBuilder = new StringBuilder(BATCH_UPDATE_FAILED);
+            for (final CompletionStage<AsyncResultSet> future : futures) {
+                try {
+                    final AsyncResultSet asyncResultSet = CompletableFutures.getUninterruptibly(future);
+                    if (asyncResultSet.getColumnDefinitions().size() > 0) {
+                        returnCounts[i] = EXECUTE_FAILED;
+                        hasFailures = true;
+                        errMsgBuilder.append(String.format(BATCH_STATEMENT_FAILURE_MSG, i,
+                            "attempts to return a result set"));
+                    } else {
+                        returnCounts[i] = SUCCESS_NO_INFO;
+                    }
+                } catch (final Exception e) {
+                    returnCounts[i] = EXECUTE_FAILED;
+                    hasFailures = true;
+                    errMsgBuilder.append(String.format(BATCH_STATEMENT_FAILURE_MSG, i, e.getMessage()));
+                } finally {
+                    i++;
+                }
+            }
+
+            // If at least one statement in the batch has failed, throw a BatchUpdateException with the appropriate
+            // updateCounts (i.e. failed statements have value EXECUTE_FAILED).
+            if (hasFailures) {
+                throw new BatchUpdateException(errMsgBuilder.toString(), returnCounts);
+            }
+        } catch (final SQLException sqlEx) {
+            // Any SQLException can be re-throw as is, without being wrapped into another SQLException.
+            throw sqlEx;
+        } catch (final Exception e) {
+            throw new SQLTransientException(e);
+        } finally {
+            // Empty the batch statement list after execution even if it failed.
+            this.batchQueries = new ArrayList<>();
         }
 
         return returnCounts;